register controller
create register controller and add register functions from user controller to register controller
Relates #8
