+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\LoginController;
+use App\Http\Controllers\RegisterController;
 use App\Http\Controllers\UserContoller;
 use Illuminate\Support\Facades\Route;
 
@@ -26,8 +27,8 @@ Route::get('/student', function () {
     return view('student');
 });
 
-Route::get('/signup', [UserContoller::class, 'signup']);
-Route::post('/signup', [UserContoller::class, 'store']);
+Route::get('/signup', [RegisterController::class, 'signup']);
+Route::post('/signup', [RegisterController::class, 'store']);
 
 Route::get('/signin', [LoginController::class, 'login']);
 Route::post('/signin', [LoginController::class, 'authenticate']);
